fix : lambda 크롤링 코드 수정 (이미지로 패키징 처리 및 성능 향상 & 기존 코드 삭제)
