+++ lambda/scrap/activity/crawling_activity.py
@@ -1,12 +1,44 @@
 from selenium import webdriver
 from selenium.webdriver.chrome.service import Service
 from selenium.webdriver.common.by import By
-from selenium.common.exceptions import NoSuchElementException, ElementClickInterceptedException, StaleElementReferenceException
+from selenium.common.exceptions import NoSuchElementException, StaleElementReferenceException
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 import os
 import json
-import boto3
+import boto3, uuid, shutil
+
+# 전역 세션 ID
+def generate_session_id():
+    return str(uuid.uuid4())
+
+def chrome_driver(session_id):
+    options = webdriver.ChromeOptions()
+    options.binary_location = "/opt/chrome/chrome"
+    options.add_argument('--no-sandbox')
+    options.add_argument("--single-process")
+    options.add_argument('--disable-gpu')
+    options.add_argument('--disable-dev-shm-usage')
+    options.add_argument("--remote-debugging-port=9222")
+    options.add_argument('--headless')
+    options.add_argument('user-agent=Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/61.0.3163.100 Safari/537.36')
+
+    options.add_argument(f"--user-data-dir=/tmp/user-data-{session_id}")
+    options.add_argument(f"--data-path=/tmp/data-path-{session_id}")
+    options.add_argument(f"--homedir=/tmp/home-{session_id}")
+    options.add_argument(f"--disk-cache-dir=/tmp/cache-dir-{session_id}")
+
+    service = Service("/opt/chromedriver")
+    driver = webdriver.Chrome(service=service, options=options)
+
+    return driver
+
+def remove_tmp(session_id):
+    shutil.rmtree(f"/tmp/user-data-{session_id}", ignore_errors=True)
+    shutil.rmtree(f"/tmp/data-path-{session_id}", ignore_errors=True)
+    shutil.rmtree(f"/tmp/home-{session_id}", ignore_errors=True)
+    shutil.rmtree(f"/tmp/cache-dir-{session_id}", ignore_errors=True)
+    
 
 # 기존의 모든 json 파일 로드
 def load_all_json_ids_from_s3(bucket_name: str, prefix: str) -> set:
@@ -26,6 +58,9 @@ def load_all_json_ids_from_s3(bucket_name: str, prefix: str) -> set:
             if key.endswith(".json"):
                 try:
                     res = s3.get_object(Bucket=bucket_name, Key=key)
+                    if res is None:
+                        continue
+
                     data = json.load(res["Body"])
                     ids = {item["linkareer_id"] for item in data if "linkareer_id" in item}
                     all_ids.update(ids)
@@ -45,69 +80,63 @@ def extract_link_id(link: str) -> int | None:
     except ValueError:
         return None
 
+
 # 각각의 link들 크롤링
-def crawling_activity_link():
-    main_url = 'https://linkareer.com/list/activity'
+def get_href(driver):
     links = []
+    page_count = 1
+    driver.get(f'https://linkareer.com/list/activity?page={page_count}')
 
     try:
-        page = 1
         while True:
-            print(f"크롤링 중: {page} 페이지")
+            print(f" === ACTIVITY 활동 크롤링 중: {page_count} 페이지 === ")
 
-            contents = WebDriverWait(driver, 10).until(
-                EC.presence_of_all_elements_located(By.CSS_SELECTOR, 'div[class^="activity-list-card-item-wrapper"]'
-                ))
+            cards = WebDriverWait(driver, 10).until(EC.presence_of_all_elements_located((By.CLASS_NAME, "ActivityListCardItem__StyledWrapper-sc-eb439f06-0")))
 
-            for content in contents:
-
+            for card in cards:
                 try:
-                    href = content.find_element(By.CSS_SELECTOR, 'a[href^="/activity/"]').get_attribute('href')
-                    full_link = "https://linkareer.com" + href if href.startswith("/") else href
-                    links.append(full_link)
+                    href = card.find_element(By.CLASS_NAME, "image-link").get_attribute('href')
+                    links.append(href)
                 except StaleElementReferenceException:
                     print("상태 불안정 에러 발생. 건너뜁니다.")
                     continue
 
-            # 다음 페이지
-            try:
-                next_page_number = str(page + 1)
-                next_page_button = driver.find_element(
-                    By.XPATH, f'//button[span[text()="{next_page_number}"]]'
-                )
-                next_page_button.click()
-                WebDriverWait(driver, 10).until(
-                    EC.presence_of_all_elements_located(By.CSS_SELECTOR, 'div[class^="activity-list-card-item-wrapper"]'
-                                                        ))
-                page += 1
+            buttons = WebDriverWait(driver, 10).until(EC.presence_of_all_elements_located((By.CSS_SELECTOR, ".button-page-number")))
 
-            except NoSuchElementException:
-                try:
-                    next_arrow = driver.find_element(By.CSS_SELECTOR, 'button.button-arrow-next')
-                    if next_arrow.get_attribute("disabled"):
-                        print("마지막 페이지입니다.")
-                        break
-                    next_arrow.click()
-                    page += 1
-                    WebDriverWait(driver, 10).until(
-                        EC.presence_of_all_elements_located(By.CSS_SELECTOR, 'div[class^="activity-list-card-item-wrapper"]'))
-                except NoSuchElementException:
-                    print("다음 버튼 없음. 종료합니다.")
-                    break
-                except ElementClickInterceptedException:
-                    print("화살표 클릭 차단됨.")
-                    break
-            except ElementClickInterceptedException:
-                print("페이지 번호 클릭 차단됨.")
-                break
+            # 현재 페이지 목록에서 최대 페이지 번호
+            max_number = 0
+            for btn in buttons:
+                max_number = max(max_number, int(btn.text.strip()))
 
+            # 현재 위치해있는 페이지와 페이지 번호
+            active_button = WebDriverWait(driver, 2).until(EC.presence_of_element_located((By.CSS_SELECTOR, ".active-page")))
+            active_number = int(active_button.text.strip())
+
+            # 다음 페이지 버튼 눌러야 할 때 (최대 페이지 번호와 현재 페이지 번호가 같을 때)
+            if(active_number == max_number):
+                next_button = driver.find_element(By.CLASS_NAME, "button-arrow-next")
+                driver.execute_script("arguments[0].scrollIntoView(true);", next_button)
+
+                # 만약 마지막 페이지 번호라 다음 목록으로 넘어가는 버튼이 활성화되지 않았다면 종료
+                if "Mui-disabled" in next_button.get_attribute("class"):
+                    return links
+
+                driver.execute_script("arguments[0].click();", next_button)
+
+            # 페이지 목록을 조회 후, 다음 눌러야할 번호가 같으면 클릭
+            buttons = driver.find_elements(By.CSS_SELECTOR, ".button-page-number")
+            for btn in buttons:
+                if btn.text.strip() == str(active_number + 1):
+                    driver.execute_script("arguments[0].click();", btn)
+                    break
+
+            # 다음 페이지 넘어간 후 페이지 번호 1 증가
+            page_count = active_number + 1
     except Exception as e:
-        print(f"크롤링 중 예외 발생: {e}")
-        traceback.print_exc()  # 추가
+        print(e)
         driver.quit()
 
-    finally:
-        return links
+    return links
 
 
 # 구체적인 내용 크롤링
@@ -123,10 +152,12 @@ def crawling_activity_detail(link: str, driver) -> dict | None:
     try:
         driver.get(link)
         WebDriverWait(driver, 10).until(
-            EC.presence_of_element_located(By.CSS_SELECTOR, 'header.ActivityInformationHeader__StyledWrapper-sc-7bdaebe9-0 h1')
-        )
+            EC.presence_of_element_located((By.CSS_SELECTOR, 'header.ActivityInformationHeader__StyledWrapper-sc-7bdaebe9-0 h1')))
+
 
         linkareer_id = extract_link_id(link)
+
+        title = driver.find_element(By.CSS_SELECTOR, "header h1").text
 
         img_url = driver.find_element(By.CSS_SELECTOR, 'img.card-image').get_attribute('src')
 
@@ -178,7 +209,7 @@ def crawling_activity_detail(link: str, driver) -> dict | None:
 
 
 # S3에 저장
-def save_activity_to_s3(bucket_name: str, driver):
+def save_activity_to_s3(bucket_name: str, driver, session_id, context):
 
     # 기존의 데이터 불러오기
     try:
@@ -189,10 +220,17 @@ def save_activity_to_s3(bucket_name: str, driver):
 
     # 새로운 크롤링
     try:
-        links = crawling_activity_link(driver)
-
+        count = 0
+        links = get_href(driver)
         new_data = []
         for link in links:
+            count += 1
+            if count % 5 == 0:
+                driver.quit()
+                remove_tmp(session_id=session_id)
+                session_id = generate_session_id()
+                driver = chrome_driver(session_id=session_id)
+
             link_id = extract_link_id(link)
             if link_id is None or link_id in old_ids:
                 print(f"이미 존재하는 공고입니다: {link_id}")
@@ -201,6 +239,10 @@ def save_activity_to_s3(bucket_name: str, driver):
             data = crawling_activity_detail(link, driver)
             if data:
                 new_data.append(data)
+                print(f"Saved: {data['title']}")
+
+            if context.get_remaining_time_in_millis() < 5000:
+                return new_data
 
         return new_data
