+++ lambda/scrap/competition/crawling_competition.py
@@ -1,13 +1,42 @@
-import traceback
 from selenium import webdriver
 from selenium.webdriver.chrome.service import Service
 from selenium.webdriver.common.by import By
 from selenium.common.exceptions import NoSuchElementException
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
-import boto3
-import os
-import json, re
+import os, shutil, uuid, json, boto3
+
+# 전역 세션 ID
+def generate_session_id():
+    return str(uuid.uuid4())
+
+def chrome_driver(session_id):
+    options = webdriver.ChromeOptions()
+    options.binary_location = "/opt/chrome/chrome"
+    options.add_argument('--no-sandbox')
+    options.add_argument("--single-process")
+    options.add_argument('--disable-gpu')
+    options.add_argument('--disable-dev-shm-usage')
+    options.add_argument("--remote-debugging-port=9222")
+    options.add_argument('--headless')
+    options.add_argument('user-agent=Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/61.0.3163.100 Safari/537.36')
+
+    options.add_argument(f"--user-data-dir=/tmp/user-data-{session_id}")
+    options.add_argument(f"--data-path=/tmp/data-path-{session_id}")
+    options.add_argument(f"--homedir=/tmp/home-{session_id}")
+    options.add_argument(f"--disk-cache-dir=/tmp/cache-dir-{session_id}")
+
+    service = Service("/opt/chromedriver")
+    driver = webdriver.Chrome(service=service, options=options)
+    
+    return driver
+
+def remove_tmp(session_id):
+    shutil.rmtree(f"/tmp/user-data-{session_id}", ignore_errors=True)
+    shutil.rmtree(f"/tmp/data-path-{session_id}", ignore_errors=True)
+    shutil.rmtree(f"/tmp/home-{session_id}", ignore_errors=True)
+    shutil.rmtree(f"/tmp/cache-dir-{session_id}", ignore_errors=True)
+    
 
 # 기존의 모든 json 파일 로드
 def load_all_json_ids_from_s3(bucket_name: str, prefix: str) -> set:
@@ -47,7 +76,6 @@ def extract_link_id(link: str) -> int | None:
         return None
 
 
-
 def safe_get_text(driver, by, value):
     try:
         return driver.find_element(by, value).text
@@ -61,24 +89,24 @@ def safe_get_attr(driver, by, value, attr):
         return '-'
 
 
-educations_dict = dict()
+competitions_dict = dict()
 
 
 def get_href(driver):
     results = []
     page_count = 1
-    driver.get(f'https://linkareer.com/list/education?page={page_count}')
+    driver.get(f'https://linkareer.com/list/contest?page={page_count}')
 
     try:
         while(True):
-            print(f" === EDUCATION 활동 크롤링 중: {page_count} 페이지 === ")
-            cards = WebDriverWait(driver, 15).until(EC.presence_of_all_elements_located((By.CLASS_NAME, "EducationListRow-desktop__StyledWrapper-sc-1c4f824c-0")))
+            print(f" === 크롤링 중: {page_count} 페이지 === ")
+            cards = WebDriverWait(driver, 15).until(EC.presence_of_all_elements_located((By.CLASS_NAME, "activity-list-card-item-wrapper")))
             # 각 ACTIVITY당 썸네일에서 파싱 가능한 데이터는 미리 dictionary에 삽입 (key, value) -> 여기서 key값은 href의 마지막 숫자 ex) 243667, value는 dict()
             for card in cards:
-                href = card.find_element(By.CLASS_NAME, "education-link").get_attribute("href")
+                href = card.find_element(By.CLASS_NAME, "image-link").get_attribute("href")
                 results.append(href)
-                education_key = href.split("/")[-1]
-                educations_dict[education_key] = dict()
+                competition_key = href.split("/")[-1]
+                competitions_dict[competition_key] = dict()
 
 
             buttons = WebDriverWait(driver, 10).until(EC.presence_of_all_elements_located((By.CSS_SELECTOR, ".button-page-number")))
@@ -114,15 +142,16 @@ def get_href(driver):
             page_count = active_number + 1
     except Exception as e:
         print(e)
-        traceback.print_exc()  # 추가
         driver.quit()
 
+    return results
 
-def save_education_to_s3(bucket_name: str, driver):
+
+def save_competition_to_s3(bucket_name: str, driver, session_id, context):
 
     # 기존의 데이터 불러오기
     try:
-        old_ids = load_all_json_ids_from_s3(bucket_name, "data/EDUCATION")
+        old_ids = load_all_json_ids_from_s3(bucket_name, "data/COMPETITION")
     except Exception as e:
         old_ids = set()
         print("파일을 읽을 수 없습니다.", e)
@@ -131,48 +160,57 @@ def save_education_to_s3(bucket_name: str, driver):
     try:
         hrefs = get_href(driver=driver)
 
+        count = 0
         for href in hrefs:
+            count += 1
+            if count % 5 == 0:
+                driver.quit()
+                remove_tmp(session_id=session_id)
+                session_id = generate_session_id()
+                driver = chrome_driver(session_id=session_id)
+
             link_id = extract_link_id(href)
             if link_id is None or link_id in old_ids:
                 print(f"이미 존재하는 공고입니다: {link_id}")
                 continue
 
 
-            print(f"progressing...{href}")
+            # 새로 크롤링
+            print(f"processing... {link_id}")
             driver.get(href)
             WebDriverWait(driver, 5).until(EC.presence_of_all_elements_located((By.CLASS_NAME, "activity-detail-content")))
 
             competition_key = href.split("/")[-1]
-            education_value = educations_dict[competition_key]
-
-            # 만약 데이터가 없을 시, pass
-            if(education_value is None):
+            competition_value = competitions_dict.get(competition_key)
+            if competition_value is None:
                 continue
 
-            education_value["organizer_name"] = safe_get_text(driver, By.XPATH, "//header[@class='organization-info']/h2[@class='organization-name']")
-            education_value["title"] = safe_get_text(driver, By.XPATH, "//header[contains(@class, 'ActivityInformationHeader__StyledWrapper')]/h1")
+            competition_value["organizer_name"] = safe_get_text(driver, By.XPATH, "//header[@class='organization-info']/h2[@class='organization-name']")
+            competition_value["title"] = safe_get_text(driver, By.XPATH, "//header[contains(@class, 'ActivityInformationHeader__StyledWrapper')]/h1")
 
-            # start_date 날짜 형식이 아닐 시 None 입력
-            start_date = safe_get_text(driver, By.XPATH, "//dl[dt[text()='접수기간']]/dd/div/span[@class='start-at']/following-sibling::span[1]").strip()
-            education_value["start_date"] = start_date
+            try:
+                competition_value["job_category"] = json.dumps([elem.text for elem in driver.find_elements(By.XPATH, "//dt[text()='공모분야']/following-sibling::dd//li/p")], ensure_ascii=False)
+            except Exception:
+                competition_value["job_category"] = json.dumps([], ensure_ascii=False)
 
-            # end_date 날짜 형식이 아닐 시 None 입력
-            end_date = safe_get_text(driver, By.XPATH, "//dl[dt[text()='접수기간']]/dd/span[@class='end-at']/following-sibling::span[1]").strip()
-            education_value["end_date"] = end_date
+            competition_value["start_date"] = safe_get_text(driver, By.XPATH, "//dl[dt[text()='접수기간']]/dd/div/span[@class='start-at']/following-sibling::span[1]")
+            competition_value["end_date"] = safe_get_text(driver, By.XPATH, "//dl[dt[text()='접수기간']]/dd/span[@class='end-at']/following-sibling::span[1]")
 
-            education_value["job_category"] = safe_get_text(driver, By.XPATH, "//dt[text()='교육분야']/following-sibling::dd[1]")
-
-            # 홈페이지 없을 시, 링커리어 링크 첨부
             homepage_url = safe_get_attr(driver, By.CSS_SELECTOR, "dd.text > a", "href")
-            if(homepage_url == '-'):
-                homepage_url = href
-            education_value["homepage_url"] = homepage_url
+            competition_value["homepage_url"] = homepage_url if homepage_url != '-' else href
 
-            education_value["img_url"] = safe_get_attr(driver, By.CLASS_NAME, "card-image", "src")
-            education_value["on_off_line"] = safe_get_text(driver, By.XPATH, "//dt[text()='온/오프라인']/following-sibling::dd[1]")
-            education_value["linkareer_id"] = competition_key
+            competition_value["img_url"] = safe_get_attr(driver, By.CLASS_NAME, "card-image", "src")
+            competition_value["benefits"] = safe_get_text(driver, By.XPATH, "//dt[text()='활동혜택']/following-sibling::dd")
+            competition_value["eligibility"] = safe_get_text(driver, By.XPATH, "//dt[text()='참여대상']/following-sibling::dd")
+            competition_value["linkareer_id"] = int(competition_key)
 
             new_data.append(competition_value)
             print(f"Saved: {competition_value['title']}")
+
+            if context.get_remaining_time_in_millis() < 5000:
+                return new_data
+
+        return new_data
     except Exception as e:
-        print(f"교육 크롤링 중 오류 발생: {e}")
+        print(f"공모전 크롤링 중 오류 발생: {e}")
+        return new_data
